Fixing a couple compiler warnings

+++ src/asar/addr2line.cpp
@@ -32,7 +32,7 @@ void AddressToLineMapping::includeMapping(const char* filename, int line, int ad
 int AddressToLineMapping::getFileIndex(const char* filename)
 {
 	// check if the file exists first
-	unsigned long filenameCrc = crc32((const unsigned char*)filename, strlen(filename));
+	unsigned long filenameCrc = crc32((const unsigned char*)filename, (unsigned int)strlen(filename));
 	for (int i = 0; i < m_filenameCrcs.count; ++i)
 	{
 		if (m_filenameCrcs[i] == filenameCrc)
